Rename setupRequestParsing to setUpRequestParsing for consistency (T #755)
